+++ src/main/java/roomescape/payment/client/TossPaymentGateway.java
@@ -6,9 +6,9 @@ import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
 import roomescape.domain.PaymentStatus;
-import roomescape.payment.PaymentConfirmation;
-import roomescape.payment.PaymentGateway;
-import roomescape.payment.PaymentResult;
+import roomescape.service.payment.PaymentConfirmation;
+import roomescape.service.payment.PaymentGateway;
+import roomescape.service.payment.PaymentResult;
 import roomescape.payment.client.dto.ConfirmRequest;
 import roomescape.payment.client.dto.TossErrorResponse;
 import roomescape.payment.client.dto.TossPaymentResponse;
