+++ src/main/java/roomescape/service/payment/PaymentAmountMismatchException.java
@@ -1,4 +1,4 @@
-package roomescape.payment;
+package roomescape.service.payment;
 
 public class PaymentAmountMismatchException extends RuntimeException {
