+++ src/main/java/roomescape/service/payment/PaymentGateway.java
@@ -1,4 +1,4 @@
-package roomescape.payment;
+package roomescape.service.payment;
 
 public interface PaymentGateway {
