+++ src/main/java/roomescape/service/payment/PaymentResult.java
@@ -1,4 +1,4 @@
-package roomescape.payment;
+package roomescape.service.payment;
 
 import roomescape.domain.PaymentStatus;
