+++ src/main/resources/application.properties
@@ -6,6 +6,6 @@ spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 
 # 학습테스트에서 제공한 Toss 테스트 키
-toss.client-key=test_gck_docs_Ovk5rk1EwkEbP0W43n07xlzm
 toss.base-url=https://api.tosspayments.com
+toss.client-key=test_gck_docs_Ovk5rk1EwkEbP0W43n07xlzm
 toss.secret-key=test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6
